Fix timeout handling in WiFiSecureConnection and improve E2E tests
- Fix timeout calculation to ensure minimum 1 second in WiFiSecureConnection
- Add debug logging for timeout in HttpClient
- Refactor E2E tests to dedicated folder test/e2e/test_real_server
- Update E2E test case to verify execution duration instead of just error code
- Update platformio.ini for better E2E test support
