+++ platformio.ini
@@ -95,12 +95,17 @@ build_flags = ${common_native.build_flags}
 extra_scripts = pre:scripts/fix_cpp_httplib.py
 
 ; 3. E2E Tests (ESP32 Real Hardware)
-[env:test_esp32]
+[env:test_e2e]
 extends = common_esp32
 test_filter = e2e/*
 test_ignore = unit/*, integration/*
 build_flags = ${common_esp32.build_flags}
               -DE2E_TEST
+# モニター設定の追加
+monitor_speed = 115200
+monitor_rts = 0
+monitor_dtr = 0
+build_src_filter = +<*> -<native/*> -<main.cpp> -<esp32_main.cpp> -<esp32/main.cpp> +<Config.cpp>
 
 ; ======= PRODUCTION ENVIRONMENTS =======
 
@@ -116,15 +121,3 @@ extends = common_esp32
 build_flags = ${common_esp32.build_flags}
               -DPRODUCTION
               -DEXTENDED_CONFIG
-
-; ======= LEGACY ENVIRONMENTS (backward compatibility) =======
-
-[env:native]
-extends = common_native
-test_filter = native_basic
-test_ignore = test_e2e_esp32.cpp, test_unit_utils.cpp, e2e/*, integration/*, unit/*
-build_flags = ${common_native.build_flags}
-              -DUNIT_TEST
-
-[env:native_integration]
-extends = env:test_integration
