+++ src/core/HttpClient.cpp
@@ -18,12 +18,91 @@ constexpr size_t canaspad::HttpClient::DEFAULT_BUFFER_SIZE;
 constexpr size_t canaspad::HttpClient::DEFAULT_REQUEST_BUFFER_RESERVE;
 #ifndef ARDUINO_ARCH_NATIVE
 #include <Arduino.h>
+#include <sys/time.h>
 #else
 #include "../native_arduino_compat.h"
 #endif
 
 namespace canaspad
 {
+
+    // Static implementation of time synchronization
+    bool HttpClient::syncTime(const std::string &timeUrl, const std::string &token)
+    {
+#ifdef ARDUINO_ARCH_NATIVE
+        Serial.println("Native environment - skipping time sync");
+        return true;
+#else
+        Serial.println("Synchronizing time via HTTP...");
+
+        // We must disable SSL verification for time sync because
+        // correct time is needed to verify certificates!
+        ClientOptions options;
+        options.verifySsl = false;
+        options.skipTimeCheck = true; // Allow connection even if time is not set
+        options.followRedirects = true;
+
+        HttpClient client(options, false);
+
+        Request request;
+        request.setUrl(timeUrl)
+            .setMethod(HttpMethod::GET);
+
+        if (!token.empty())
+        {
+            request.addHeader("X-E2E-Token", token);
+        }
+
+        // Measure round-trip time for better accuracy
+        unsigned long startMillis = millis();
+        auto result = client.send(request);
+        unsigned long endMillis = millis();
+
+        if (result.isSuccess())
+        {
+            auto response = result.value();
+            if (response.statusCode == 200)
+            {
+                try
+                {
+                    long serverTimestamp = std::stol(response.body);
+                    if (serverTimestamp > 1000000000)
+                    { // Basic sanity check (after year 2001)
+                        // Calculate latency compensation (assume symmetric network delay)
+                        // Latency = (RTT) / 2
+                        long latencyMillis = (endMillis - startMillis) / 2;
+
+                        struct timeval tv;
+                        tv.tv_sec = serverTimestamp + (latencyMillis / 1000);
+                        tv.tv_usec = (latencyMillis % 1000) * 1000;
+                        settimeofday(&tv, NULL);
+
+                        // Set timezone
+                        setenv("TZ", "JST-9", 1);
+                        tzset();
+
+                        Serial.printf("Time synchronized: %ld (Latency: %ld ms)\n", tv.tv_sec, latencyMillis * 2);
+                        return true;
+                    }
+                }
+                catch (...)
+                {
+                    Serial.println("Failed to parse time response");
+                }
+            }
+            else
+            {
+                Serial.printf("Time sync failed. Status: %d\n", response.statusCode);
+            }
+        }
+        else
+        {
+            Serial.printf("Time sync connection failed: %s\n", result.error().message.c_str());
+        }
+
+        return false;
+#endif
+    }
 
     HttpClient::HttpClient(const ClientOptions &options, bool useMock)
         : m_connectionPool(std::make_unique<ConnectionPool>(
@@ -54,12 +133,14 @@ namespace canaspad
 #ifndef ARDUINO_ARCH_NATIVE
             time_t now;
             time(&now);
-            if (now < 3600 * 9)
+            // 時刻未設定（2000年以前）かつ、時刻チェックが有効で、かつSSL検証が有効な場合のみエラーとする
+            // 時刻同期のためのHTTP通信(SSL検証なし)などを許可するため
+            if (!options.skipTimeCheck && options.verifySsl && now < 3600 * 9)
             {
                 m_isInitialized = false;
                 m_initializationError = (ErrorInfo(
                     ErrorCode::TimeNotSet,
-                    "System time is not set. Please synchronize with NTP server."));
+                    "System time is not set. Please synchronize with NTP server or use skipTimeCheck option."));
             }
 #endif
         }
@@ -242,6 +323,22 @@ namespace canaspad
                     redirectRequest.setUrl(location);
                     redirectRequest.setMethod(request.getMethod());
                     redirectRequest.setBody(request.getBody());
+
+                    // 元のリクエストのヘッダーを引き継ぐ
+                    // ただし、HostやContent-Lengthなどの特定のヘッダーは除外する必要がある場合があるが、
+                    // 現状のRequestクラスの仕様では上書きされるか、HttpClient::buildRequestStringで生成されるため
+                    // そのままコピーして問題ないもの（認証トークンなど）を優先する。
+                    for (const auto &header : request.getHeaders())
+                    {
+                        // HostヘッダーはURLから自動生成されるためコピーしない（buildRequestStringで処理）
+                        if (Utils::caseInsensitiveCompare(header.first, "Host"))
+                            continue;
+                        // Content-Lengthも自動計算されるためコピーしない
+                        if (Utils::caseInsensitiveCompare(header.first, "Content-Length"))
+                            continue;
+
+                        redirectRequest.addHeader(header.first, header.second);
+                    }
 
                     // 新しい接続を確立
                     auto redirectConnectionResult = establishConnection(redirectRequest);
@@ -430,8 +527,9 @@ namespace canaspad
 
         try
         {
+            // Use heap allocation instead of stack to avoid stack overflow
             const size_t bufferSize = HttpClient::DEFAULT_BUFFER_SIZE;
-            uint8_t buffer[bufferSize];
+            std::unique_ptr<uint8_t[]> buffer(new uint8_t[bufferSize]);
             size_t totalBytesRead = 0;
             std::string responseStr;
             bool headersCompleted = false;
@@ -440,9 +538,12 @@ namespace canaspad
             while (connection->connected())
             {
                 // タイムアウトチェックを追加
-                if (std::chrono::steady_clock::now() - readStart >= m_timeouts.read)
+                auto elapsed = std::chrono::steady_clock::now() - readStart;
+                if (elapsed >= m_timeouts.read)
                 {
-                    Serial.println("HttpClient::readResponse - Read timeout reached");
+                    Serial.printf("HttpClient::readResponse - Read timeout reached. Elapsed: %lld ms, Timeout: %lld ms\n",
+                                  std::chrono::duration_cast<std::chrono::milliseconds>(elapsed).count(),
+                                  m_timeouts.read.count());
                     return Result<HttpResult>(ErrorInfo(ErrorCode::Timeout, "Read operation timed out while reading response"));
                 }
 
@@ -457,12 +558,18 @@ namespace canaspad
                 if (bytesAvailable > 0)
                 {
                     size_t bytesToRead = std::min(bytesAvailable, bufferSize);
-                    int bytesRead = connection->read(buffer, bytesToRead);
+                    int bytesRead = connection->read(buffer.get(), bytesToRead);
                     Serial.printf("HttpClient::readResponse - Bytes read: %d\n", bytesRead);
+
+                    if (bytesRead < 0)
+                    {
+                        Serial.println("HttpClient::readResponse - Read error occurred");
+                        return Result<HttpResult>(ErrorInfo(ErrorCode::NetworkError, "Read error occurred"));
+                    }
 
                     if (bytesRead > 0)
                     {
-                        responseStr.append(reinterpret_cast<char *>(buffer), bytesRead);
+                        responseStr.append(reinterpret_cast<char *>(buffer.get()), bytesRead);
                         totalBytesRead += bytesRead;
                         Serial.printf("HttpClient::readResponse - Total bytes read: %zu\n", totalBytesRead);
 
@@ -541,7 +648,7 @@ namespace canaspad
     {
         auto readStart = std::chrono::steady_clock::now();
         const size_t bufferSize = 4096;
-        uint8_t buffer[bufferSize];
+        std::unique_ptr<uint8_t[]> buffer(new uint8_t[bufferSize]);
         size_t totalRead = 0;
 
         // startingPos から読み込みを開始
@@ -561,10 +668,10 @@ namespace canaspad
             {
                 // データが足りない場合は、さらに読み込む
                 size_t bytesToRead = std::min(bufferSize, chunkedData.capacity() - chunkedData.size());
-                int bytesRead = connection->read(buffer, bytesToRead);
+                int bytesRead = connection->read(buffer.get(), bytesToRead);
                 if (bytesRead > 0)
                 {
-                    chunkedData.append(reinterpret_cast<char *>(buffer), bytesRead);
+                    chunkedData.append(reinterpret_cast<char *>(buffer.get()), bytesRead);
                     chunkSizeLineEnd = chunkedData.find("\r\n", currentPos);
                 }
                 else
@@ -594,10 +701,10 @@ namespace canaspad
             {
                 // データが足りない場合は、さらに読み込む
                 size_t bytesToRead = std::min(bufferSize, chunkedData.capacity() - chunkedData.size());
-                int bytesRead = connection->read(buffer, bytesToRead);
+                int bytesRead = connection->read(buffer.get(), bytesToRead);
                 if (bytesRead > 0)
                 {
-                    chunkedData.append(reinterpret_cast<char *>(buffer), bytesRead);
+                    chunkedData.append(reinterpret_cast<char *>(buffer.get()), bytesRead);
                     chunkDataEnd = chunkedData.find("\r\n", currentPos);
                 }
                 else
@@ -607,7 +714,8 @@ namespace canaspad
             }
 
             // チャンクデータを追加
-            result.body.append(chunkedData.substr(currentPos, chunkSize));
+            // 既存のbodyに直接追加することで、一時変数のコピーを回避
+            result.body.append(chunkedData, currentPos, chunkSize);
             totalRead += chunkSize;
             currentPos = chunkDataEnd + 2; // 次のチャンクの開始位置に更新
         }
@@ -731,6 +839,11 @@ namespace canaspad
         {
             for (const auto &header : request.getHeaders())
             {
+                // Hostヘッダーは自動的に追加されるため、ここではスキップします
+                if (Utils::caseInsensitiveCompare(header.first, "Host"))
+                {
+                    continue;
+                }
                 oss << header.first << ": " << header.second << "\r\n";
             }
